Consume the new UI ZIP assembly from backend.
This updates the backend project to unzip the content of the UI assembly
and then copy it to the WEB-INF/static directory when the WAR generation
happens.  The more wide open resource mapping is once again needed.

At this point a full build from the parent POM creates a WAR that embeds
the UI resources.