Fix handling of pure dates between DB and UI.
Hibernate (or JDBC) was treating dates as a full time when retrieving from
the database, which resulted in a time of 00:00:00 UTC.  Which was then
converted to the JVM timezone, changing the date to the previous day.

Extending TimestampType is a technique I used before, and the type just
wasn't being recognized in this case.  Rather than waste more time on it,
I opted to try out @Converter.