Add a bunch of new functionality and refactor.
This is really an example of what happens when you lose track of things.

This change introduces sub modules for the different UI components and
child route management.  Some common data pipes are included.

Also adds a bunch of model objects to both the UI and the backend
projects.

UI and backend changes are added that relate to the following new
functionality: programs, payroll, and service companies.